Potential fix for code scanning alert no. 1: Multiplication result converted to larger type (#267)

+++ src/AT_Algorithms_GSM.c
@@ -655,7 +655,7 @@ void AT_GSM_multiple_local_dose_distrib( const long  number_of_field_components,
 		}
 
 		*mean_d_check_Gy		+=	cur_d_check_Gy;
-		*sd_d_check_Gy			+=	cur_d_check_Gy * cur_d_check_Gy;
+		*sd_d_check_Gy			+=	(double)cur_d_check_Gy * cur_d_check_Gy;
 	}
 
 	/* Effective calculation of running mean and stdev of x in n runs: */
